+++ packages/mochi-rest/src/modules/module.ts
@@ -1,0 +1,6 @@
+import { QueryStringAddon } from "wretch/addons/queryString";
+import { Wretch } from "wretch/types";
+
+export class Module {
+  constructor(public api: QueryStringAddon & Wretch<QueryStringAddon>) {}
+}